go-echo-backend/deployment/tf/modules/lambda/classify/requirements.txt
numpy~=1.26.4
pillow~=10.2.0
onnxruntime~=1.17.1
requests~=2.31.0